Move embed into Yggdrasil demo
