+++ cmd/dendrite-demo-yggdrasil/embed/embed_other.go
@@ -1,0 +1,7 @@
+// +build !riotweb
+
+package embed
+
+func Embed(_ int, _ string) {
+
+}
