--- internal/embed/embed_other.go
@@ -1,7 +1,0 @@
-// +build !riotweb
-
-package embed
-
-func Embed(_ int, _ string) {
-
-}
